requirements.txt
pytorch
torchvision
matplotlib